backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic-settings
passlib[bcrypt]
python-dotenv
